+++ src/argilla_server/contexts/questions.py
@@ -12,72 +12,26 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from typing import List, Union
+from typing import Union
 from uuid import UUID
 
-from sqlalchemy import func, select
+from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 
 import argilla_server.errors.future as errors
-from argilla_server.enums import QuestionType
 from argilla_server.models import Dataset, Question, User
 from argilla_server.policies import QuestionPolicyV1, authorize
 from argilla_server.schemas.v1.questions import (
-    LabelSelectionQuestionSettings,
-    LabelSelectionSettingsUpdate,
+
     QuestionCreate,
-    QuestionSettings,
-    QuestionSettingsUpdate,
     QuestionUpdate,
-    SpanQuestionSettingsCreate,
 )
-
-
-class InvalidQuestionSettings(Exception):
-    pass
-
-
-def _validate_settings_type(settings: QuestionSettings, settings_update: QuestionSettingsUpdate):
-    if settings.type != settings_update.type:
-        raise InvalidQuestionSettings(
-            f"Question type cannot be changed. Expected '{settings.type}' but got '{settings_update.type}'"
-        )
-
-
-def _validate_label_options(settings: LabelSelectionQuestionSettings, settings_update: LabelSelectionSettingsUpdate):
-    # TODO: Validate visible_options on update
-    if settings_update.options is None:
-        return
-
-    if len(settings.options) != len(settings_update.options):
-        raise InvalidQuestionSettings(
-            "The number of options cannot be modified. "
-            f"Expected {len(settings.options)} but got {len(settings_update.options)}"
-        )
-
-    sorted_options = sorted(settings.options, key=lambda option: option.value)
-    sorted_update_options = sorted(settings_update.options, key=lambda option: option.value)
-
-    unexpected_options: List[str] = []
-    for option, update_option in zip(sorted_options, sorted_update_options):
-        if option.value != update_option.value:
-            unexpected_options.append(update_option.value)
-
-    if unexpected_options:
-        raise InvalidQuestionSettings(
-            f"The option values cannot be modified. " f"Found unexpected option values: {unexpected_options!r}"
-        )
-
-
-def _validate_question_settings(
-    question_settings: QuestionSettings,
-    question_update_settings: QuestionSettingsUpdate,
-) -> None:
-    _validate_settings_type(question_settings, question_update_settings)
-
-    if question_settings.type in [QuestionType.label_selection, QuestionType.multi_label_selection, QuestionType.span]:
-        _validate_label_options(question_settings, question_update_settings)
+from argilla_server.validators.questions import (
+    QuestionCreateValidator,
+    QuestionDeleteValidator,
+    QuestionUpdateValidator,
+)
 
 
 async def get_question_by_id(db: AsyncSession, question_id: UUID) -> Union[Question, None]:
@@ -98,35 +52,8 @@ async def get_question_by_name_and_dataset_id_or_raise(db: AsyncSession, name: s
     return question
 
 
-def _validate_question_before_create(dataset: Dataset, question_create: QuestionCreate) -> None:
-    if question_create.settings.type == QuestionType.span:
-        _validate_span_question_settings_before_create(dataset, question_create.settings)
-
-
-def _validate_question_before_update(question: Question, question_update: QuestionUpdate) -> None:
-    if question_update.settings:
-        _validate_question_settings(question.parsed_settings, question_update.settings)
-
-
-def _validate_span_question_settings_before_create(
-    dataset: Dataset, span_question_settings_create: SpanQuestionSettingsCreate
-) -> None:
-    field = span_question_settings_create.field
-    field_names = [field.name for field in dataset.fields]
-
-    if field not in field_names:
-        raise ValueError(f"'{field}' is not a valid field name.\nValid field names are {field_names!r}")
-
-    for question in dataset.questions:
-        if question.type == QuestionType.span and field == question.parsed_settings.field:
-            raise ValueError(f"'{field}' is already used by span question with id '{question.id}'")
-
-
 async def create_question(db: AsyncSession, dataset: Dataset, question_create: QuestionCreate) -> Question:
-    if dataset.is_ready:
-        raise ValueError("Question cannot be created for a published dataset")
-
-    _validate_question_before_create(dataset, question_create)
+    QuestionCreateValidator(question_create).validate_for(dataset)
 
     return await Question.create(
         db,
@@ -148,14 +75,13 @@ async def update_question(
 
     await authorize(current_user, QuestionPolicyV1.update(question))
 
-    _validate_question_before_update(question, question_update)
+    QuestionUpdateValidator(question_update).validate_for(question)
 
     params = question_update.dict(exclude_unset=True)
     return await question.update(db, **params)
 
 
 async def delete_question(db: AsyncSession, question: Question) -> Question:
-    if question.dataset.is_ready:
-        raise ValueError("Questions cannot be deleted for a published dataset")
+    QuestionDeleteValidator().validate_for(question.dataset)
 
     return await question.delete(db)